requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.2
python-dotenv==1.0.1
pydantic==2.9.2
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
